requirements.txt
Pysocks
colored
undetected_chromedriver
